Move abc_renderer as well.
